Rename distributor files
