+++ src/pytket_dqc/distributors/__init__.py
@@ -1,10 +1,10 @@
 from .distributor import Distributor  # noqa:F401
-from .bipartite_embedding import (  # noqa:F401
+from .cover_embedding import (  # noqa:F401
     CoverEmbedding,
     CoverEmbeddingSteiner,
     CoverEmbeddingSteinerDetached,
 )
-from .partitioning_embedding import (  # noqa:F401
+from .partitioning_heterogeneous import (  # noqa:F401
     PartitioningHeterogeneous,
     PartitioningHeterogeneousEmbedding,
 )
